+++ Content.Server/AlertLevel/AlertLevelSystem.cs
@@ -5,6 +5,7 @@ using Content.Server.Station.Systems;
 using Robust.Shared.Audio;
 using Robust.Shared.Player;
 using Robust.Shared.Prototypes;
+using Content.Shared.AlertLevel;
 
 namespace Content.Server.AlertLevel;
 
@@ -198,14 +199,3 @@ public sealed class AlertLevelDelayFinishedEvent : EntityEventArgs
 public sealed class AlertLevelPrototypeReloadedEvent : EntityEventArgs
 {}
 
-public sealed class AlertLevelChangedEvent : EntityEventArgs
-{
-    public EntityUid Station { get; }
-    public string AlertLevel { get; }
-
-    public AlertLevelChangedEvent(EntityUid station, string alertLevel)
-    {
-        Station = station;
-        AlertLevel = alertLevel;
-    }
-}
