+++ Content.Server/Communications/CommunicationsConsoleSystem.cs
@@ -12,6 +12,7 @@ using Content.Server.Shuttles.Systems;
 using Content.Server.Station.Systems;
 using Content.Shared.Access.Components;
 using Content.Shared.Access.Systems;
+using Content.Shared.AlertLevel;
 using Content.Shared.CCVar;
 using Content.Shared.Communications;
 using Content.Shared.Database;
@@ -105,29 +106,6 @@ namespace Content.Server.Communications
                 if (args.Station == entStation)
                 {
                     UpdateCommsConsoleInterface(comp);
-                }
-            }
-
-            // Gives an access to appropriate departments to doors of other departments in case of emergency 
-            foreach (var (_, access) in EntityQuery<AirlockComponent, AccessReaderBoardComponent>(true))
-            {
-                if (TryComp(access.BoardContainer.ContainedEntities.FirstOrDefault(), out AccessStorageComponent? accessStorage)
-                     && IsAirlockSupportAlertLevel(accessStorage.AccessLists))
-                {
-                    switch (args.AlertLevel)
-                    {
-                        case "yellow":
-                            ClearAlertLevelAccess(accessStorage);
-                            accessStorage.AccessLists.Add(new HashSet<string>() { "Engineering" });
-                            break;
-                        case "red":
-                            ClearAlertLevelAccess(accessStorage);
-                            accessStorage.AccessLists.Add(new HashSet<string>() { "Brig" });
-                            break;
-                        default:
-                            ClearAlertLevelAccess(accessStorage);
-                            break;
-                    }
                 }
             }
         }
@@ -331,31 +309,6 @@ namespace Content.Server.Communications
 
             _roundEndSystem.CancelRoundEndCountdown(uid);
             _adminLogger.Add(LogType.Action, LogImpact.Extreme, $"{ToPrettyString(mob):player} has recalled the shuttle.");
-        }
-
-        private bool IsAirlockSupportAlertLevel (List<HashSet<string>> accessLists)
-        {
-            var exceptions = new List<string>() { "Captain",
-                    "HeadOfPersonnel",
-                    "HeadOfSecurity",
-                    "Quartermaster",
-                    "ResearchDirector",
-                    "ChiefEngineer",
-                    "ChiefMedicalOfficer",
-                    "Command",
-                    "Armory"
-                };
-            return accessLists.Count > 0 && !accessLists[0].Any(a => exceptions.Contains(a));
-        }
-
-
-
-        private void ClearAlertLevelAccess(AccessStorageComponent accessStorage)
-        {
-            if (accessStorage.AccessLists.ElementAtOrDefault(1) != null)
-            {
-                accessStorage.AccessLists.RemoveAt(1);
-            }
         }
     }
 }
