+++ Content.Server/Light/EntitySystems/EmergencyLightSystem.cs
@@ -4,6 +4,7 @@ using Content.Server.Light.Components;
 using Content.Server.Power.Components;
 using Content.Server.Station.Components;
 using Content.Server.Station.Systems;
+using Content.Shared.AlertLevel;
 using Content.Shared.Examine;
 using Content.Shared.Light;
 using Content.Shared.Light.Component;
